refactor: split oversized test monoliths into focused route-family modules
